Split postProcessing into smaller modules
