+++ lib/jsonApi.js
@@ -1,5 +1,7 @@
 "use strict";
 var jsonApi = module.exports = { };
+jsonApi._resources = { };
+jsonApi._apiConfig = { };
 
 var _ = require("underscore");
 var Joi = require("joi");
@@ -7,19 +9,14 @@ var router = require("./router.js");
 var responseHelper = require("./responseHelper.js");
 var routes = require("./routes");
 var mockHandlers = require("./mockHandlers.js");
-var postProcess = require("./postProcess.js");
 
 jsonApi.mockHandlers = mockHandlers.handlers;
-jsonApi._resources = { };
-jsonApi._apiConfig = { };
 
 jsonApi.setConfig = function(apiConfig) {
   jsonApi._apiConfig = apiConfig;
   jsonApi._apiConfig.base = jsonApi._cleanBaseUrl(jsonApi._apiConfig.base);
   responseHelper.setBaseUrl(jsonApi._apiConfig.base);
   responseHelper.setMetadata(jsonApi._apiConfig.meta);
-  router.using(jsonApi);
-  postProcess.using(jsonApi);
 };
 
 jsonApi._cleanBaseUrl = function(base) {
