+++ lib/router.js
@@ -6,6 +6,7 @@ var express = require("express");
 var app = express();
 var server;
 var bodyParser = require("body-parser");
+var jsonApi = require("./jsonApi.js");
 
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
@@ -31,14 +32,10 @@ router.close = function() {
   server = null;
 };
 
-router.using = function(jsonApi) {
-  router._jsonApi = jsonApi;
-};
-
 router.bindRoute = function(config, callback) {
-  app[config.verb](router._jsonApi._apiConfig.base + config.path, function(req, res) {
+  app[config.verb](jsonApi._apiConfig.base + config.path, function(req, res) {
     var request = router._getParams(req);
-    var resourceConfig = router._jsonApi._resources[request.params.type];
+    var resourceConfig = jsonApi._resources[request.params.type];
     router._setResponseHeaders(request, res);
     request.resourceConfig = resourceConfig;
     return callback(request, resourceConfig, res);
@@ -68,7 +65,7 @@ router._getParams = function(req) {
     headers: req.headers,
     route: {
       host: req.headers.host,
-      base: router._jsonApi._apiConfig.base,
+      base: jsonApi._apiConfig.base,
       path: urlParts.shift() || "",
       query: urlParts.shift() || "",
       combined: "https://" + req.headers.host + req.url
