+++ lib/routes/related.js
@@ -1,6 +1,7 @@
 "use strict";
 var relatedRoute = module.exports = { };
 
+var jsonApi = require("../jsonApi.js");
 var async = require("async");
 var helper = require("./helper.js");
 var router = require("../router.js");
@@ -46,7 +47,7 @@ relatedRoute.register = function() {
         if (relation._settings.__one) {
           relatedResources = relatedResources[0];
         }
-        request.resourceConfig = router._jsonApi._resources[relation._settings.__one || relation._settings.__many];
+        request.resourceConfig = jsonApi._resources[relation._settings.__one || relation._settings.__many];
         response = responseHelper._generateResponse(request, resourceConfig, relatedResources);
         response.included = [ ];
         postProcess.handle(request, response, callback);
